revamp of installation scripts to more intuitively determine order
auto kernel building is now the by-default. it is possible
to still use manual_kernel_build instead of 40_build_kernel
but since the kernel is already supposed to be generic enough
to run on all hardware, not going to recommend it as an option
unless the user already has the notion to manually build

also assuming that the installation will always delete the
partitions existing as well as redoing the fs. considering the
throwaway nature of qubes--and the speediness of full re-do's--
it's recommended to just do ./1, ./2, ./3 ... anytime it needs
to be restarted, rather than attempt to just re-run a single
script and program in idompotence